// src/sim/data/info.gen.ts
export const SPR = {
  TROO: 0,
  SHTG: 1,
  PUNG: 2,
  PISG: 3,
  PISF: 4,
  SHTF: 5,
  SHT2: 6,
  CHGG: 7,
  CHGF: 8,
  MISG: 9,
  MISF: 10,
  SAWG: 11,
  PLSG: 12,
  PLSF: 13,
  BFGG: 14,
  BFGF: 15,
  BLUD: 16,
  PUFF: 17,
  BAL1: 18,
  BAL2: 19,
  PLSS: 20,
  PLSE: 21,
  MISL: 22,
  BFS1: 23,
  BFE1: 24,
  BFE2: 25,
  TFOG: 26,
  IFOG: 27,
  PLAY: 28,
  POSS: 29,
  SPOS: 30,
  VILE: 31,
  FIRE: 32,
  FATB: 33,
  FBXP: 34,
  SKEL: 35,
  MANF: 36,
  FATT: 37,
  CPOS: 38,
  SARG: 39,
  HEAD: 40,
  BAL7: 41,
  BOSS: 42,
  BOS2: 43,
  SKUL: 44,
  SPID: 45,
  BSPI: 46,
  APLS: 47,
  APBX: 48,
  CYBR: 49,
  PAIN: 50,
  SSWV: 51,
  KEEN: 52,
  BBRN: 53,
  BOSF: 54,
  ARM1: 55,
  ARM2: 56,
  BAR1: 57,
  BEXP: 58,
  FCAN: 59,
  BON1: 60,
  BON2: 61,
  BKEY: 62,
  RKEY: 63,
  YKEY: 64,
  BSKU: 65,
  RSKU: 66,
  YSKU: 67,
  STIM: 68,
  MEDI: 69,
  SOUL: 70,
  PINV: 71,
  PSTR: 72,
  PINS: 73,
  MEGA: 74,
  SUIT: 75,
  PMAP: 76,
  PVIS: 77,
  CLIP: 78,
  AMMO: 79,
  ROCK: 80,
  BROK: 81,
  CELL: 82,
  CELP: 83,
  SHEL: 84,
  SBOX: 85,
  BPAK: 86,
  BFUG: 87,
  MGUN: 88,
  CSAW: 89,
  LAUN: 90,
  PLAS: 91,
  SHOT: 92,
  SGN2: 93,
  COLU: 94,
  SMT2: 95,
  GOR1: 96,
  POL2: 97,
  POL5: 98,
  POL4: 99,
  POL3: 100,
  POL1: 101,
  POL6: 102,
  GOR2: 103,
  GOR3: 104,
  GOR4: 105,
  GOR5: 106,
  SMIT: 107,
  COL1: 108,
  COL2: 109,
  COL3: 110,
  COL4: 111,
  CAND: 112,
  CBRA: 113,
  COL6: 114,
  TRE1: 115,
  TRE2: 116,
  ELEC: 117,
  CEYE: 118,
  FSKU: 119,
  COL5: 120,
  TBLU: 121,
  TGRN: 122,
  TRED: 123,
  SMBT: 124,
  SMGT: 125,
  SMRT: 126,
  HDB1: 127,
  HDB2: 128,
  HDB3: 129,
  HDB4: 130,
  HDB5: 131,
  HDB6: 132,
  POB1: 133,
  POB2: 134,
  BRS1: 135,
  TLMP: 136,
  TLP2: 137,
  NUMSPRITES: 138,
} as const;
// ...
export const MT = {
  PLAYER: 0,
  POSSESSED: 1,
  SHOTGUY: 2,
  VILE: 3,
  FIRE: 4,
  UNDEAD: 5,
  TRACER: 6,
  SMOKE: 7,
  FATSO: 8,
  FATSHOT: 9,
  CHAINGUY: 10,
  TROOP: 11,
  SERGEANT: 12,
  SHADOWS: 13,
  HEAD: 14,
  BRUISER: 15,
  BRUISERSHOT: 16,
  KNIGHT: 17,
  SKULL: 18,
  SPIDER: 19,
  BABY: 20,
  CYBORG: 21,
  PAIN: 22,
  WOLFSS: 23,
  KEEN: 24,
  BOSSBRAIN: 25,
  BOSSSPIT: 26,
  BOSSTARGET: 27,
  SPAWNSHOT: 28,
  SPAWNFIRE: 29,
  BARREL: 30,
  TROOPSHOT: 31,
  HEADSHOT: 32,
  ROCKET: 33,
  PLASMA: 34,
  BFG: 35,
  ARACHPLAZ: 36,
  PUFF: 37,
  BLOOD: 38,
  TFOG: 39,
  IFOG: 40,
  TELEPORTMAN: 41,
  EXTRABFG: 42,
  MISC0: 43,
  MISC1: 44,
  MISC2: 45,
  MISC3: 46,
  MISC4: 47,
  MISC5: 48,
  MISC6: 49,
  MISC7: 50,
  MISC8: 51,
  MISC9: 52,
  MISC10: 53,
  MISC11: 54,
  MISC12: 55,
  INV: 56,
  MISC13: 57,
  INS: 58,
  MISC14: 59,
  MISC15: 60,
  MISC16: 61,
  MEGA: 62,
  CLIP: 63,
  MISC17: 64,
  MISC18: 65,
  MISC19: 66,
  MISC20: 67,
  MISC21: 68,
  MISC22: 69,
  MISC23: 70,
  MISC24: 71,
  MISC25: 72,
  CHAINGUN: 73,
  MISC26: 74,
  MISC27: 75,
  MISC28: 76,
  SHOTGUN: 77,
  SUPERSHOTGUN: 78,
  MISC29: 79,
  MISC30: 80,
  MISC31: 81,
  MISC32: 82,
  MISC33: 83,
  MISC34: 84,
  MISC35: 85,
  MISC36: 86,
  MISC37: 87,
  MISC38: 88,
  MISC39: 89,
  MISC40: 90,
  MISC41: 91,
  MISC42: 92,
  MISC43: 93,
  MISC44: 94,
  MISC45: 95,
  MISC46: 96,
  MISC47: 97,
  MISC48: 98,
  MISC49: 99,
  MISC50: 100,
  MISC51: 101,
  MISC52: 102,
  MISC53: 103,
  MISC54: 104,
  MISC55: 105,
  MISC56: 106,
  MISC57: 107,
  MISC58: 108,
  MISC59: 109,
  MISC60: 110,
  MISC61: 111,
  MISC62: 112,
  MISC63: 113,
  MISC64: 114,
  MISC65: 115,
  MISC66: 116,
  MISC67: 117,
  MISC68: 118,
  MISC69: 119,
  MISC70: 120,
  MISC71: 121,
  MISC72: 122,
  MISC73: 123,
  MISC74: 124,
  MISC75: 125,
  MISC76: 126,
  MISC77: 127,
  MISC78: 128,
  MISC79: 129,
  MISC80: 130,
  MISC81: 131,
  MISC82: 132,
  MISC83: 133,
  MISC84: 134,
  MISC85: 135,
  MISC86: 136,
  NUMMOBJTYPES: 137,
} as const;

// mobj flags (prefix MF_ stripped).
export const MF = {
  SPECIAL: 1,
  SOLID: 2,
  SHOOTABLE: 4,
  NOSECTOR: 8,
  NOBLOCKMAP: 16,
  AMBUSH: 32,
  JUSTHIT: 64,
  JUSTATTACKED: 128,
  SPAWNCEILING: 256,
  NOGRAVITY: 512,
  DROPOFF: 1024,
  PICKUP: 2048,
  NOCLIP: 4096,
  SLIDE: 8192,
  FLOAT: 16384,
  TELEPORT: 32768,
  MISSILE: 65536,
  DROPPED: 131072,
  SHADOW: 262144,
  NOBLOOD: 524288,
  CORPSE: 1048576,
  INFLOAT: 2097152,
  COUNTKILL: 4194304,
  COUNTITEM: 8388608,
  SKULLFLY: 16777216,
  NOTDMATCH: 33554432,
  TRANSLATION: 201326592,
  TRANSSHIFT: 26,
} as const;

export const sprnames: readonly string[] = ["TROO","SHTG","PUNG","PISG","PISF","SHTF","SHT2","CHGG","CHGF","MISG","MISF","SAWG","PLSG","PLSF","BFGG","BFGF","BLUD","PUFF","BAL1","BAL2","PLSS","PLSE","MISL","BFS1","BFE1","BFE2","TFOG","IFOG","PLAY","POSS","SPOS","VILE","FIRE","FATB","FBXP","SKEL","MANF","FATT","CPOS","SARG","HEAD","BAL7","BOSS","BOS2","SKUL","SPID","BSPI","APLS","APBX","CYBR","PAIN","SSWV","KEEN","BBRN","BOSF","ARM1","ARM2","BAR1","BEXP","FCAN","BON1","BON2","BKEY","RKEY","YKEY","BSKU","RSKU","YSKU","STIM","MEDI","SOUL","PINV","PSTR","PINS","MEGA","SUIT","PMAP","PVIS","CLIP","AMMO","ROCK","BROK","CELL","CELP","SHEL","SBOX","BPAK","BFUG","MGUN","CSAW","LAUN","PLAS","SHOT","SGN2","COLU","SMT2","GOR1","POL2","POL5","POL4","POL3","POL1","POL6","GOR2","GOR3","GOR4","GOR5","SMIT","COL1","COL2","COL3","COL4","CAND","CBRA","COL6","TRE1","TRE2","ELEC","CEYE","FSKU","COL5","TBLU","TGRN","TRED","SMBT","SMGT","SMRT","HDB1","HDB2","HDB3","HDB4","HDB5","HDB6","POB1","POB2","BRS1","TLMP","TLP2"];