// src/sim/data/sounds.gen.ts
export const SFX = {
  None: 0,
  pistol: 1,
  shotgn: 2,
  sgcock: 3,
  dshtgn: 4,
  dbopn: 5,
  dbcls: 6,
  dbload: 7,
  plasma: 8,
  bfg: 9,
  sawup: 10,
  sawidl: 11,
  sawful: 12,
  sawhit: 13,
  rlaunc: 14,
  rxplod: 15,
  firsht: 16,
  firxpl: 17,
  pstart: 18,
  pstop: 19,
  doropn: 20,
  dorcls: 21,
  stnmov: 22,
  swtchn: 23,
  swtchx: 24,
  plpain: 25,
  dmpain: 26,
  popain: 27,
  vipain: 28,
  mnpain: 29,
  pepain: 30,
  slop: 31,
  itemup: 32,
  wpnup: 33,
  oof: 34,
  telept: 35,
  posit1: 36,
  posit2: 37,
  posit3: 38,
  bgsit1: 39,
  bgsit2: 40,
  sgtsit: 41,
  cacsit: 42,
  brssit: 43,
  cybsit: 44,
  spisit: 45,
  bspsit: 46,
  kntsit: 47,
  vilsit: 48,
  mansit: 49,
  pesit: 50,
  sklatk: 51,
  sgtatk: 52,
  skepch: 53,
  vilatk: 54,
  claw: 55,
  skeswg: 56,
  pldeth: 57,
  pdiehi: 58,
  podth1: 59,
  podth2: 60,
  podth3: 61,
  bgdth1: 62,
  bgdth2: 63,
  sgtdth: 64,
  cacdth: 65,
  skldth: 66,
  brsdth: 67,
  cybdth: 68,
  spidth: 69,
  bspdth: 70,
  vildth: 71,
  kntdth: 72,
  pedth: 73,
  skedth: 74,
  posact: 75,
  bgact: 76,
  dmact: 77,
  bspact: 78,
  bspwlk: 79,
  vilact: 80,
  noway: 81,
  barexp: 82,
  punch: 83,
  hoof: 84,
  metal: 85,
  chgun: 86,
  tink: 87,
  bdopn: 88,
  bdcls: 89,
  itmbk: 90,
  flame: 91,
  flamst: 92,
  getpow: 93,
  bospit: 94,
  boscub: 95,
  bossit: 96,
  bospn: 97,
  bosdth: 98,
  manatk: 99,
  mandth: 100,
  sssit: 101,
  ssdth: 102,
  keenpn: 103,
  keendt: 104,
  skeact: 105,
  skesit: 106,
  skeatk: 107,
  radio: 108,
  NUMSFX: 109,
} as const;